Beginnings of the hub system todo: comb over for all cases where it would rather use hub than area
